+++ RELEASE-NOTES.md
@@ -34,6 +34,8 @@
 - The `incomplete_beta` function in `pymc3.distributions.dist_math` was replaced by `aesara.tensor.betainc` (see [4857](https://github.com/pymc-devs/pymc3/pull/4857)).
 - `math.log1mexp` and `math.log1mexp_numpy` will expect negative inputs in the future. A `FutureWarning` is now raised unless `negative_input=True` is set (see [#4860](https://github.com/pymc-devs/pymc3/pull/4860)).
 - Change name of `Lognormal` distribution to `LogNormal` to harmonize CamelCase usage for distribution names.
+- Attempt to iterate over MultiTrace will raise NotImplementedError
+- ...
 
 ## PyMC3 3.11.2 (14 March 2021)